Added method to set solution by player
I added a method to check for human_codebreaker when getting a solution.
If false (=> CODEBREAKER = Human) the code is generated by Computer (run
Computer.getting_solution).
If true, Player.getting_solution is being called which will set solution
based on Player input.
However, code as is throws error and seems to collide with some of the
methods in the system.